requirements.txt
requests>=2.32.5
discord.py>=2.6.3
aiosqlite>=0.21.0
unidecode>=1.4.0